requirements.txt
redis
PyYAML
